Move pkg/testutils and pkg/jsonschema into internal
