+++ internal/jsonschema/testdata/add-column-1.txtar
