+++ internal/jsonschema/testdata/alter-column-1.txtar
