+++ internal/jsonschema/testdata/alter-column-2.txtar
