+++ internal/jsonschema/testdata/alter-column-3.txtar
