+++ internal/jsonschema/testdata/alter-column-4.txtar
