+++ internal/jsonschema/testdata/alter-column-5.txtar
