+++ internal/jsonschema/testdata/alter-column-6.txtar
