+++ internal/jsonschema/testdata/alter-column-7.txtar
