+++ internal/jsonschema/testdata/alter-column-8.txtar
