+++ internal/jsonschema/testdata/alter-column-9.txtar
