+++ internal/jsonschema/testdata/create-table-1.txtar
