+++ internal/jsonschema/testdata/rename-constraint-1.txtar
