+++ internal/jsonschema/testdata/sql-1.txtar
