+++ internal/jsonschema/testdata/sql-2.txtar
